refactor: rename nestjs files, token AI_GATEWAY_CLIENT→AI_GATEWAY_CONFIG, trim peerDeps (phase 04)
- interfaces.ts → ai-gateway.types.ts
- constants.ts → ai-gateway.constants.ts (AI_GATEWAY_CLIENT → AI_GATEWAY_CONFIG)
- decorators.ts → ai-gateway.decorators.ts
- forRootAsync now strips isGlobal before providing config (consistent with forRoot)
- AI_GATEWAY_OPTIONS removed from public barrel exports (internal wiring detail)
- peerDependencies: removed ai and @ai-sdk/provider; keep @ai-sdk/openai + @nestjs/common
