+++ package.json
@@ -109,18 +109,10 @@
   },
   "peerDependencies": {
     "@ai-sdk/openai": "^3.0.0",
-    "@ai-sdk/provider": "^3.0.0",
-    "@nestjs/common": ">=10.0.0",
-    "ai": "^6.0.0"
+    "@nestjs/common": ">=10.0.0"
   },
   "peerDependenciesMeta": {
-    "ai": {
-      "optional": true
-    },
     "@ai-sdk/openai": {
-      "optional": true
-    },
-    "@ai-sdk/provider": {
       "optional": true
     },
     "@nestjs/common": {
