+++ src/nestjs/__tests__/decorators.spec.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { InjectAIGateway } from '../decorators';
+import { InjectAIGateway } from '../ai-gateway.decorators';
 
 describe('InjectAIGateway', () => {
   it('returns a decorator function', () => {
