+++ src/nestjs/ai-gateway.constants.ts
@@ -1,5 +1,5 @@
 /** Injection token for the GatewayProviderSettings object. */
-export const AI_GATEWAY_CLIENT = Symbol('AI_GATEWAY_CLIENT');
+export const AI_GATEWAY_CONFIG = Symbol('AI_GATEWAY_CONFIG');
 
 /** Injection token for the module config options. */
 export const AI_GATEWAY_OPTIONS = Symbol('AI_GATEWAY_OPTIONS');
