+++ src/nestjs/index.ts
@@ -1,5 +1,5 @@
 export { AIGatewayModule } from './ai-gateway.module';
 export { AIGatewayExceptionFilter } from './ai-gateway.filter';
-export { InjectAIGateway } from './decorators';
-export { AI_GATEWAY_CLIENT, AI_GATEWAY_OPTIONS } from './constants';
-export type { AIGatewayModuleOptions, AIGatewayModuleAsyncOptions, AIGatewayOptionsFactory } from './interfaces';
+export { InjectAIGateway } from './ai-gateway.decorators';
+export { AI_GATEWAY_CONFIG } from './ai-gateway.constants';
+export type { AIGatewayModuleOptions, AIGatewayModuleAsyncOptions, AIGatewayOptionsFactory } from './ai-gateway.types';
